Crate google_androiddeviceprovisioning1

This documentation was generated from Android Provisioning Partner crate version 1.0.9+20190622, where 20190622 is the exact revision of the androiddeviceprovisioning:v1 schema built by the mako code generator v1.0.9.

Everything else about the Android Provisioning Partner v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

customers

configurations create, configurations delete, configurations get, configurations list, configurations patch, devices apply configuration, devices get, devices list, devices remove configuration, devices unclaim, dpcs list and list

operations

get

partners

customers create, customers list, devices claim, devices claim async, devices find by identifier, devices find by owner, devices get, devices metadata, devices unclaim, devices unclaim async, devices update metadata async, vendors customers list and vendors list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.customers().devices_apply_configuration(...).doit()
let r = hub.customers().devices_unclaim(...).doit()
let r = hub.customers().configurations_delete(...).doit()
let r = hub.customers().devices_remove_configuration(...).doit()
let r = hub.partners().devices_unclaim(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-androiddeviceprovisioning1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androiddeviceprovisioning1 as androiddeviceprovisioning1;
use androiddeviceprovisioning1::CustomerApplyConfigurationRequest;
use androiddeviceprovisioning1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androiddeviceprovisioning1::AndroidProvisioningPartner;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidProvisioningPartner::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = CustomerApplyConfigurationRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.customers().devices_apply_configuration(req, "parent")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AndroidProvisioningPartner
Central instance to access all AndroidProvisioningPartner related resource activities

Chunk

ClaimDeviceRequest
Request message to claim a device on behalf of a customer.

ClaimDeviceResponse
Response message containing device id of the claim.

ClaimDevicesRequest
Request to claim devices asynchronously in batch. Claiming a device adds the device to zero-touch enrollment and shows the device in the customer's view of the portal.

Company
A reseller, vendor, or customer in the zero-touch reseller and customer APIs.

Configuration
A configuration collects the provisioning options for Android devices. Each configuration combines the following:

ContentRange
Implements the Content-Range header, for serialization only

CreateCustomerRequest
Request message to create a customer.

CustomerApplyConfigurationRequest
Request message for customer to assign a configuration to device.

CustomerConfigurationCreateCall
Creates a new configuration. Once created, a customer can apply the configuration to devices.

CustomerConfigurationDeleteCall
Deletes an unused configuration. The API call fails if the customer has devices with the configuration applied.

CustomerConfigurationGetCall
Gets the details of a configuration.

CustomerConfigurationListCall
Lists a customer's configurations.

CustomerConfigurationPatchCall
Updates a configuration's field values.

CustomerDeviceApplyConfigurationCall
Applies a Configuration to the device to register the device for zero-touch enrollment. After applying a configuration to a device, the device automatically provisions itself on first boot, or next factory reset.

CustomerDeviceGetCall
Gets the details of a device.

CustomerDeviceListCall
Lists a customer's devices.

CustomerDeviceRemoveConfigurationCall
Removes a configuration from device.

CustomerDeviceUnclaimCall
Unclaims a device from a customer and removes it from zero-touch enrollment.

CustomerDpcListCall
Lists the DPCs (device policy controllers) that support zero-touch enrollment.

CustomerListCall
Lists the user's customer accounts.

CustomerListConfigurationsResponse
Response message of customer's listing configuration.

CustomerListCustomersResponse
Response message for listing my customers.

CustomerListDevicesResponse
Response message of customer's liting devices.

CustomerListDpcsResponse
Response message of customer's listing DPCs.

CustomerMethods
A builder providing access to all methods supported on customer resources. It is not used directly, but through the AndroidProvisioningPartner hub.

CustomerRemoveConfigurationRequest
Request message for customer to remove the configuration from device.

CustomerUnclaimDeviceRequest
Request message for customer to unclaim a device.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Device
An Android device registered for zero-touch enrollment.

DeviceClaim
A record of a device claimed by a reseller for a customer. Devices claimed for zero-touch enrollment have a claim with the type SECTION_TYPE_ZERO_TOUCH. To learn more, read Claim devices for customers.

DeviceIdentifier
Encapsulates hardware and product IDs to identify a manufactured device. To understand requirements on identifier sets, read Identifiers.

DeviceMetadata
Metadata entries that can be attached to a Device. To learn more, read Device metadata.

DeviceReference
A DeviceReference is an API abstraction that lets you supply a device argument to a method using one of the following identifier types:

Dpc
An EMM's DPC (device policy controller). Zero-touch enrollment installs a DPC (listed in the Configuration) on a device to maintain the customer's mobile policies. All the DPCs listed by the API support zero-touch enrollment and are available in Google Play.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

FindDevicesByDeviceIdentifierRequest
Request to find devices.

FindDevicesByDeviceIdentifierResponse
Response containing found devices.

FindDevicesByOwnerRequest
Request to find devices by customers.

FindDevicesByOwnerResponse
Response containing found devices.

JsonServerError
A utility type which can decode a server response that indicates error

ListCustomersResponse
Response message of all customers related to this partner.

ListVendorCustomersResponse
Response message to list customers of the vendor.

ListVendorsResponse
Response message to list vendors of the partner.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Operation
This resource represents a long-running operation that is the result of a network API call.

OperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

OperationMethods
A builder providing access to all methods supported on operation resources. It is not used directly, but through the AndroidProvisioningPartner hub.

PartnerClaim
Identifies one claim request.

PartnerCustomerCreateCall
Creates a customer for zero-touch enrollment. After the method returns successfully, admin and owner roles can manage devices and EMM configs by calling API methods or using their zero-touch enrollment portal. The customer receives an email that welcomes them to zero-touch enrollment and explains how to sign into the portal.

PartnerCustomerListCall
Lists the customers that are enrolled to the reseller identified by the partnerId argument. This list includes customers that the reseller created and customers that enrolled themselves using the portal.

PartnerDeviceClaimAsyncCall
Claims a batch of devices for a customer asynchronously. Adds the devices to zero-touch enrollment. To learn more, read Long‑running batch operations.

PartnerDeviceClaimCall
Claims a device for a customer and adds it to zero-touch enrollment. If the device is already claimed by another customer, the call returns an error.

PartnerDeviceFindByIdentifierCall
Finds devices by hardware identifiers, such as IMEI.

PartnerDeviceFindByOwnerCall
Finds devices claimed for customers. The results only contain devices registered to the reseller that's identified by the partnerId argument. The customer's devices purchased from other resellers don't appear in the results.

PartnerDeviceGetCall
Gets a device.

PartnerDeviceMetadataCall
Updates reseller metadata associated with the device.

PartnerDeviceUnclaimAsyncCall
Unclaims a batch of devices for a customer asynchronously. Removes the devices from zero-touch enrollment. To learn more, read Long‑running batch operations.

PartnerDeviceUnclaimCall
Unclaims a device from a customer and removes it from zero-touch enrollment.

PartnerDeviceUpdateMetadataAsyncCall
Updates the reseller metadata attached to a batch of devices. This method updates devices asynchronously and returns an Operation that can be used to track progress. Read Long‑running batch operations.

PartnerMethods
A builder providing access to all methods supported on partner resources. It is not used directly, but through the AndroidProvisioningPartner hub.

PartnerUnclaim
Identifies one unclaim request.

PartnerVendorCustomerListCall
Lists the customers of the vendor.

PartnerVendorListCall
Lists the vendors of the partner.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

UnclaimDeviceRequest
Request message to unclaim a device.

UnclaimDevicesRequest
Request to unclaim devices asynchronously in batch.

UpdateDeviceMetadataInBatchRequest
Request to update device metadata in batch.

UpdateDeviceMetadataRequest
Request to set metadata for a device.

UpdateMetadataArguments
Identifies metdata updates to one device.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.