Struct google_androiddeviceprovisioning1::PartnerMethods

pub struct PartnerMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on partner resources. It is not used directly, but through the AndroidProvisioningPartner hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androiddeviceprovisioning1 as androiddeviceprovisioning1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androiddeviceprovisioning1::AndroidProvisioningPartner;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidProvisioningPartner::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `customers_create(...)`, `customers_list(...)`, `devices_claim(...)`, `devices_claim_async(...)`, `devices_find_by_identifier(...)`, `devices_find_by_owner(...)`, `devices_get(...)`, `devices_metadata(...)`, `devices_unclaim(...)`, `devices_unclaim_async(...)`, `devices_update_metadata_async(...)`, `vendors_customers_list(...)` and `vendors_list(...)`
// to build up your call.
let rb = hub.partners();

Methods

impl<'a, C, A> PartnerMethods<'a, C, A>

pub fn devices_update_metadata_async(
&self,
request: UpdateDeviceMetadataInBatchRequest,
partner_id: &str
) -> PartnerDeviceUpdateMetadataAsyncCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the reseller metadata attached to a batch of devices. This method updates devices asynchronously and returns an Operation that can be used to track progress. Read Long‑running batch operations.

Arguments

request - No description provided.

partnerId - Required. The reseller partner ID.

pub fn devices_get(&self, name: &str) -> PartnerDeviceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a device.

Arguments

name - Required. The device API resource name in the format partners/[PARTNER_ID]/devices/[DEVICE_ID].

pub fn devices_find_by_owner(
&self,
request: FindDevicesByOwnerRequest,
partner_id: &str
) -> PartnerDeviceFindByOwnerCall<'a, C, A>

Create a builder to help you perform the following task:

Finds devices claimed for customers. The results only contain devices registered to the reseller that's identified by the partnerId argument. The customer's devices purchased from other resellers don't appear in the results.

Arguments

request - No description provided.

partnerId - Required. The ID of the reseller partner.

pub fn customers_list(
&self,
partner_id: &str
) -> PartnerCustomerListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the customers that are enrolled to the reseller identified by the partnerId argument. This list includes customers that the reseller created and customers that enrolled themselves using the portal.

Arguments

partnerId - Required. The ID of the reseller partner.

pub fn devices_find_by_identifier(
&self,
request: FindDevicesByDeviceIdentifierRequest,
partner_id: &str
) -> PartnerDeviceFindByIdentifierCall<'a, C, A>

Create a builder to help you perform the following task:

Finds devices by hardware identifiers, such as IMEI.

Arguments

request - No description provided.

partnerId - Required. The ID of the reseller partner.

pub fn vendors_customers_list(
&self,
parent: &str
) -> PartnerVendorCustomerListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the customers of the vendor.

Arguments

parent - Required. The resource name in the format partners/[PARTNER_ID]/vendors/[VENDOR_ID].

pub fn vendors_list(&self, parent: &str) -> PartnerVendorListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the vendors of the partner.

Arguments

parent - Required. The resource name in the format partners/[PARTNER_ID].

pub fn devices_claim_async(
&self,
request: ClaimDevicesRequest,
partner_id: &str
) -> PartnerDeviceClaimAsyncCall<'a, C, A>

Create a builder to help you perform the following task:

Claims a batch of devices for a customer asynchronously. Adds the devices to zero-touch enrollment. To learn more, read Long‑running batch operations.

Arguments

request - No description provided.

partnerId - Required. The ID of the reseller partner.

pub fn devices_claim(
&self,
request: ClaimDeviceRequest,
partner_id: &str
) -> PartnerDeviceClaimCall<'a, C, A>

Create a builder to help you perform the following task:

Claims a device for a customer and adds it to zero-touch enrollment. If the device is already claimed by another customer, the call returns an error.

Arguments

request - No description provided.

partnerId - Required. The ID of the reseller partner.

pub fn devices_unclaim_async(
&self,
request: UnclaimDevicesRequest,
partner_id: &str
) -> PartnerDeviceUnclaimAsyncCall<'a, C, A>

Create a builder to help you perform the following task:

Unclaims a batch of devices for a customer asynchronously. Removes the devices from zero-touch enrollment. To learn more, read Long‑running batch operations.

Arguments

request - No description provided.

partnerId - Required. The reseller partner ID.

pub fn devices_metadata(
&self,
request: UpdateDeviceMetadataRequest,
metadata_owner_id: &str,
device_id: &str
) -> PartnerDeviceMetadataCall<'a, C, A>

Create a builder to help you perform the following task:

Updates reseller metadata associated with the device.

Arguments

request - No description provided.

metadataOwnerId - Required. The owner of the newly set metadata. Set this to the partner ID.

deviceId - Required. The ID of the device.

pub fn customers_create(
&self,
request: CreateCustomerRequest,
parent: &str
) -> PartnerCustomerCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a customer for zero-touch enrollment. After the method returns successfully, admin and owner roles can manage devices and EMM configs by calling API methods or using their zero-touch enrollment portal. The customer receives an email that welcomes them to zero-touch enrollment and explains how to sign into the portal.

Arguments

request - No description provided.

parent - Required. The parent resource ID in the format partners/[PARTNER_ID] that identifies the reseller.

pub fn devices_unclaim(
&self,
request: UnclaimDeviceRequest,
partner_id: &str
) -> PartnerDeviceUnclaimCall<'a, C, A>

Create a builder to help you perform the following task:

Unclaims a device from a customer and removes it from zero-touch enrollment.

Arguments

request - No description provided.

partnerId - Required. The ID of the reseller partner.